requirements.txt
fastapi
uvicorn
pandas
numpy
pulp
python-multipart
google-generativeai